+++ src/Type/PollOption.php
@@ -17,9 +17,13 @@ final readonly class PollOption
      * @param MessageEntity[]|null $textEntities
      */
     public function __construct(
+        public string $persistentId,
         public string $text,
         public int $voterCount,
         #[ArrayOfObjectsValue(MessageEntity::class)]
         public ?array $textEntities = null,
+        public ?User $addedByUser = null,
+        public ?Chat $addedByChat = null,
+        public ?int $additionDate = null,
     ) {}
 }